Refactor how exclusion/inclusion of pre-commit hooks is managed.
Instead of having 2 flags (optional and excludable) to control how a rule is managed, use an enumeration.
For now, the enumaration specify 3 types of rules; mandatory, opt-out, opt-in.

Mandatory rules, as the name say, cannot be switched off.
opt-out rules as included by default but can be switched off with.
opt-in rules as excluded by default and can be included.